+++ typescript/agentkit/src/action-providers/index.ts
@@ -12,3 +12,4 @@ export * from "./farcaster";
 export * from "./twitter";
 export * from "./wallet";
 export * from "./customActionProvider";
+export * from "./alchemy";